Refactor EditCard.tsx to follow SOLID principles
This refactoring improves code organization and maintainability by extracting
components and hooks into focused, single-responsibility modules.

Changes:
- Extract AuthenticationGate component (authentication logic)
- Extract business logic hooks (useOnRemove, useOnColorChange, useSaveAnnotation)
- Extract ActionButtons components (SaveButton, CancelButton)
- Extract AnnotationEditor component (annotation textarea)
- Add comprehensive comments throughout explaining intent and complexity

Benefits:
- Main EditCard.tsx reduced from 527 to 424 lines (-19.5%)
- Each extracted module has one clear responsibility
- Improved testability (modules can be tested in isolation)
- Enhanced reusability (hooks and components usable elsewhere)
- Better maintainability through clear separation of concerns
- Comprehensive JSDoc comments help future developers

SOLID Principles Applied:
- Single Responsibility: Each file handles one specific concern
- Open/Closed: Extracted hooks easier to extend without modifying core
- Interface Segregation: Minimal, focused prop interfaces

No breaking changes - maintains existing API and behavior.

Related to CORE-1453 and CORE-1452
