+++ src/app/content/highlights/components/EditCard.tsx
@@ -1,140 +1,184 @@
+/**
+ * EditCard Component
+ *
+ * Provides an editable interface for highlight annotations in REX.
+ *
+ * ## Responsibilities
+ *
+ * - **Authentication**: Prompts login for unauthenticated users
+ * - **Highlight Creation**: Creates new highlights from text selections
+ * - **Highlight Editing**: Allows color changes and annotation updates
+ * - **Highlight Deletion**: Removes highlights with confirmation
+ *
+ * ## Component Architecture
+ *
+ * The EditCard has been refactored to follow SOLID principles by extracting
+ * components and hooks into separate modules:
+ *
+ * - `EditCard/AuthenticationGate.tsx` - Authentication logic
+ * - `EditCard/hooks.ts` - Business logic (remove, color change, save)
+ * - `EditCard/ActionButtons.tsx` - Save and Cancel buttons
+ * - `EditCard/AnnotationEditor.tsx` - Annotation textarea
+ *
+ * ## State Management Strategy
+ *
+ * **Local State** (this component):
+ * - `pendingAnnotation` - Current text in textarea (may differ from saved)
+ * - `editingAnnotation` - Whether annotation textarea is in edit mode
+ * - `confirmingDelete` - Whether delete confirmation dialog is showing
+ *
+ * **Redux State** (via actions/selectors):
+ * - Saved highlight data (color, annotation, position)
+ * - User authentication status
+ * - Annotation changes pending flag
+ *
+ * ## Event Flow
+ *
+ * 1. User selects text → Highlight created → EditCard shown
+ * 2. User chooses color → useOnColorChange hook → Redux update
+ * 3. User types note → AnnotationEditor → Local state update
+ * 4. User clicks Save → SaveButton → useSaveAnnotation → Redux update
+ * 5. User clicks outside → Focus lost → Card hidden
+ *
+ * @example
+ * <EditCard
+ *   isActive={true}
+ *   highlight={highlight}
+ *   pageId="page-123"
+ *   locationFilterId="chapter-5"
+ *   onCreate={handleCreate}
+ *   onBlur={handleBlur}
+ *   // ... other props
+ * />
+ */
+
 import { Highlight } from '@openstax/highlighter';
-import { HighlightColorEnum } from '@openstax/highlighter/dist/api';
-import { HTMLElement, HTMLTextAreaElement, FocusEvent } from '@openstax/types/lib.dom';
+import { FocusEvent, HTMLElement } from '@openstax/types/lib.dom';
 import React from 'react';
-import { FormattedMessage, useIntl } from 'react-intl';
-import { useDispatch, useSelector } from 'react-redux';
+import { useDispatch } from 'react-redux';
 import styled, { css } from 'styled-components/macro';
 import { useAnalyticsEvent } from '../../../../helpers/analytics';
-import * as selectAuth from '../../../auth/selectors';
-import Button, { ButtonGroup } from '../../../components/Button';
-import { useFocusElement, useOnEsc, useTrapTabNavigation } from '../../../reactUtils';
+import { ButtonGroup } from '../../../components/Button';
+import { useTrapTabNavigation } from '../../../reactUtils';
 import theme from '../../../theme';
-import { assertDefined, assertWindow, mergeRefs } from '../../../utils';
-import { highlightStyles } from '../../constants';
-import { cardWidth } from '../constants';
-import defer from 'lodash/fp/defer';
+import { MAIN_CONTENT_ID } from '../../../context/constants';
 import {
   clearFocusedHighlight,
   setAnnotationChangesPending as setAnnotationChangesPendingAction,
-  updateHighlight,
 } from '../actions';
 import { cardPadding } from '../constants';
 import { HighlightData } from '../types';
-import { generateUpdatePayload } from './cardUtils';
 import ColorPicker from './ColorPicker';
 import Confirmation from './Confirmation';
-import Note from './Note';
-import scrollHighlightIntoView from './utils/scrollHighlightIntoView';
-import { MAIN_CONTENT_ID } from '../../../context/constants';
+import { LoginOrEdit } from './EditCard/AuthenticationGate';
+import { SaveButton, CancelButton } from './EditCard/ActionButtons';
+import { AnnotationEditor } from './EditCard/AnnotationEditor';
+import { useOnRemove, useOnColorChange, useSaveAnnotation } from './EditCard/hooks';
 
+/**
+ * Props for EditCard component
+ */
 export interface EditCardProps {
+  /** Whether the card is currently active/visible */
   isActive: boolean;
+  /** Whether there are unsaved annotation changes */
   hasUnsavedHighlight: boolean;
+  /** The highlight object being edited */
   highlight: Highlight;
+  /** Filter ID for the current location (e.g., chapter) */
   locationFilterId: string;
+  /** ID of the current page */
   pageId: string;
+  /** Callback when highlight is created */
   onCreate: (isDefaultColor: boolean) => void;
+  /** Callback when focus leaves the card */
   onBlur: typeof clearFocusedHighlight;
+  /** Callback to set annotation changes pending flag */
   setAnnotationChangesPending: typeof setAnnotationChangesPendingAction;
+  /** Callback to remove the highlight */
   onRemove: () => void;
+  /** Callback when editing is canceled */
   onCancel: () => void;
+  /** Callback when card height changes */
   onHeightChange: (ref: React.RefObject<HTMLElement>) => void;
+  /** Saved highlight data (undefined for new highlights) */
   data?: HighlightData;
+  /** CSS className for styled component */
   className: string;
+  /** Whether the card should auto-focus */
   shouldFocusCard: boolean;
+  /** Whether to show minimized version (not implemented) */
   minimize?: boolean;
 }
 
+/**
+ * EditCard - Root component for highlight editing interface
+ *
+ * This is a simple wrapper that:
+ * 1. Returns null when not active (card is hidden)
+ * 2. Forwards ref to inner components
+ * 3. Delegates to LoginOrEdit for authentication check
+ *
+ * The forwarded ref allows parent components to measure and position
+ * the card relative to the highlight.
+ */
+// tslint:disable-next-line:variable-name
 const EditCard = React.forwardRef<HTMLElement, EditCardProps>((props, ref) => {
+  const element = React.useRef<HTMLElement>(null);
+
   if (!props.isActive) {
     return null;
   }
-  return <LoginOrEdit props={props} fref={ref} />;
-});
 
-function LoginOrEdit({
-  props,
-  fref,
-}: {
-  props: React.PropsWithChildren<EditCardProps>;
-  fref: React.ForwardedRef<HTMLElement>;
-}) {
-  const authenticated = !!useSelector(selectAuth.user);
-  const element = React.useRef<HTMLElement>(null);
-  const {formatMessage} = useIntl();
-  const showCard = React.useCallback((event: React.MouseEvent) => {
-    if (event.button === 0) {
-      event.preventDefault();
-      document?.dispatchEvent(new CustomEvent('showCardEvent', { bubbles: true }));
-    }
-  }, []);
   const isNewSelection = props.highlight.elements.length === 0;
 
   return (
-    <div
+    <LoginOrEdit
       className={props.className}
-      role='dialog'
-      aria-label={formatMessage({id: 'i18n:highlighter:edit-note:label'})}
+      isNewSelection={isNewSelection}
+      shouldFocusCard={props.shouldFocusCard}
+      hasAnnotation={Boolean(props.data?.annotation)}
+      onBlur={props.onBlur}
+      fref={ref}
+      elementRef={element}
     >
-      {
-        authenticated ? (
-          <HiddenOnMobile>
-            {
-              (props.shouldFocusCard || props.data?.annotation) ? (
-                <form
-                  ref={mergeRefs(fref, element)}
-                  data-analytics-region='edit-note'
-                  data-highlight-card
-                >
-                  <ActiveEditCard props={props} element={element} />
-                </form>
-              ) :
-              <button type='button' onMouseDown={showCard}>
-                <FormattedMessage id={
-                  isNewSelection
-                    ? 'i18n:highlighting:create-instructions'
-                    : 'i18n:highlighting:instructions'
-                } />
-              </button>
-            }
-          </HiddenOnMobile>
-        ) : <LoginConfirmation onBlur={props.onBlur} />
-      }
-    </div>
+      <ActiveEditCard props={props} element={element} />
+    </LoginOrEdit>
   );
-}
+});
 
-function LoginConfirmation({
-  onBlur,
-}: Pick<EditCardProps, 'data' | 'onBlur'>) {
-  const loginLink = useSelector(selectAuth.loginLink);
-  const trackShowLogin = useAnalyticsEvent('showLogin');
-
-  React.useEffect(() => {
-    trackShowLogin();
-  }, [trackShowLogin]);
-
-  return (
-    <Confirmation
-      data-analytics-label='login'
-      data-analytics-region='highlighting-login'
-      message='i18n:highlighting:login:prompt'
-      confirmMessage='i18n:highlighting:login:link'
-      confirmLink={loginLink}
-      onCancel={onBlur}
-      drawFocus={false}
-    />
-  );
-}
-
-const HiddenOnMobile = styled.div`
-  min-width: ${cardWidth}rem;
-  ${theme.breakpoints.touchDeviceQuery(css`
-    display: none;
-  `)}
-`;
-
+/**
+ * ActiveEditCard - Main editing interface component
+ *
+ * This component coordinates the editing interface, managing:
+ * - Annotation text state (pending vs saved)
+ * - Editing mode state
+ * - Delete confirmation state
+ * - Focus and blur behavior
+ * - Analytics tracking
+ *
+ * ## State Management
+ *
+ * The component maintains three pieces of local state:
+ *
+ * 1. **pendingAnnotation**: Current text in textarea, may differ from saved value
+ * 2. **editingAnnotation**: Whether textarea is in edit mode (shows Save/Cancel)
+ * 3. **confirmingDelete**: Whether delete confirmation dialog is showing
+ *
+ * ## Focus Management
+ *
+ * The card needs careful focus management to work correctly:
+ *
+ * - When clicking outside the card → Blur card (unless editing)
+ * - When clicking in main content → Remove text selection
+ * - When pressing Escape → Cancel editing (handled by CancelButton)
+ * - When pressing Tab → Trap focus within card (accessibility)
+ *
+ * ## Height Change Notifications
+ *
+ * The card notifies its parent when its height changes (e.g., when
+ * showing/hiding buttons) so the parent can reposition it correctly.
+ */
 function ActiveEditCard({
   props,
   element,
@@ -142,35 +186,58 @@ function ActiveEditCard({
   props: React.PropsWithChildren<EditCardProps>;
   element: React.RefObject<HTMLElement>;
 }) {
-  const defaultAnnotation = React.useMemo(() => props.data?.annotation ?? '', [
-    props.data,
-  ]);
-  const [pendingAnnotation, setPendingAnnotation] = React.useState<string>(
-    defaultAnnotation
-  );
+  // ========================================
+  // Local State
+  // ========================================
+
+  const defaultAnnotation = React.useMemo(() => props.data?.annotation ?? '', [props.data]);
+
+  const [pendingAnnotation, setPendingAnnotation] = React.useState<string>(defaultAnnotation);
+
   const resetAnnotation = React.useCallback(() => {
     setPendingAnnotation(defaultAnnotation);
   }, [defaultAnnotation]);
-  const [editingAnnotation, setEditing] = React.useState(
-    Boolean(props?.data?.annotation)
-  );
-  const [confirmingDelete, setConfirmingDelete] = React.useState<boolean>(
-    false
-  );
+
+  const [editingAnnotation, setEditing] = React.useState(Boolean(props?.data?.annotation));
+
+  const [confirmingDelete, setConfirmingDelete] = React.useState<boolean>(false);
+
+  // ========================================
+  // Focus and Blur Handling
+  // ========================================
 
   const { onBlur, hasUnsavedHighlight } = props;
-  // Focus the highlight when clicking outside the Card in order to keep navigation in order
+
+  /**
+   * Blur handler that only triggers when not editing
+   *
+   * We don't want to blur the card if the user is actively editing
+   * or if there are unsaved changes, as that would lose their work.
+   */
   const blurIfNotEditing = React.useCallback(() => {
     if (!hasUnsavedHighlight && !editingAnnotation) {
       onBlur();
     }
   }, [onBlur, hasUnsavedHighlight, editingAnnotation]);
 
+  /**
+   * Handles focus moving to main content area
+   *
+   * When focus moves from the card to the main content, we need to:
+   * 1. Blur the card (if not editing)
+   * 2. Remove the text selection (if highlight not yet created)
+   *
+   * This prevents the selection from persisting after the user
+   * clicks away without creating a highlight.
+   */
   const deselectRange = React.useCallback(
-    ({target}: FocusEvent) => {
+    ({ target }: FocusEvent) => {
       const targetAsNode = target as HTMLElement;
       const mainEl = document?.getElementById(MAIN_CONTENT_ID);
 
+      // Only deselect if:
+      // - Highlight hasn't been created yet (!props.data?.color)
+      // - Focus moved to main content area
       if (!props.data?.color && targetAsNode !== mainEl && mainEl?.contains(targetAsNode)) {
         blurIfNotEditing();
         document?.getSelection()?.removeAllRanges();
@@ -179,14 +246,22 @@ function ActiveEditCard({
     [blurIfNotEditing, props.data?.color]
   );
 
-  React.useEffect(
-    () => {
-      document?.addEventListener('focusin', deselectRange);
-      return () => document?.removeEventListener('focusin', deselectRange);
-    },
-    [deselectRange]
-  );
+  // Listen for focus changes to handle deselection
+  React.useEffect(() => {
+    document?.addEventListener('focusin', deselectRange);
+    return () => document?.removeEventListener('focusin', deselectRange);
+  }, [deselectRange]);
 
+  // ========================================
+  // Height Change Notifications
+  // ========================================
+
+  /**
+   * Notify parent when card element is available
+   *
+   * The parent needs the element ref to calculate positioning
+   * and respond to height changes.
+   */
   const onHeightChange = props.onHeightChange;
   React.useEffect(() => {
     if (element.current) {
@@ -194,34 +269,98 @@ function ActiveEditCard({
     }
   }, [element, onHeightChange]);
 
+  // ========================================
+  // Analytics Tracking
+  // ========================================
+
   const trackShowCreate = useAnalyticsEvent('showCreate');
 
+  /**
+   * Track analytics when showing create card
+   *
+   * We only track this event for new highlights (props.data is undefined).
+   * This helps us measure how many highlight creation flows are started.
+   */
   React.useEffect(() => {
     if (!props.data) {
       trackShowCreate();
     }
   }, [props.data, trackShowCreate]);
 
-  const onColorChange = useOnColorChange(props);
-  const saveAnnotation = useSaveAnnotation(props, element, pendingAnnotation);
+  // ========================================
+  // Business Logic Hooks
+  // ========================================
+
+  const onColorChange = useOnColorChange({
+    highlight: props.highlight,
+    data: props.data,
+    locationFilterId: props.locationFilterId,
+    pageId: props.pageId,
+    onCreate: props.onCreate,
+  });
+
+  const saveAnnotation = useSaveAnnotation(
+    {
+      data: props.data,
+      pageId: props.pageId,
+      locationFilterId: props.locationFilterId,
+      highlight: props.highlight,
+      onCancel: props.onCancel,
+    },
+    element,
+    pendingAnnotation
+  );
+
+  const removeHighlight = useOnRemove(
+    {
+      data: props.data,
+      onRemove: props.onRemove,
+    },
+    pendingAnnotation
+  );
+
+  // ========================================
+  // Tab Navigation (Accessibility)
+  // ========================================
+
   const ref = React.useRef<HTMLElement>(null);
 
+  /**
+   * Trap tab navigation when editing
+   *
+   * When the annotation textarea is in edit mode, trap focus within
+   * the card so keyboard users can't accidentally tab out and lose focus.
+   * This is important for accessibility.
+   */
   useTrapTabNavigation(ref, editingAnnotation);
+
+  // ========================================
+  // Render
+  // ========================================
 
   return (
     <div ref={ref}>
+      {/* Color Picker */}
       <ColorPicker
         color={props.data?.color}
         onChange={onColorChange}
-        onRemove={useOnRemove(props, pendingAnnotation)}
+        onRemove={removeHighlight}
       />
+
+      {/* Annotation Editor */}
       <AnnotationEditor
-        props={props}
+        highlight={props.highlight}
+        data={props.data}
         pendingAnnotation={pendingAnnotation}
-        onColorChange={onColorChange}
+        hasUnsavedHighlight={props.hasUnsavedHighlight}
+        shouldFocusCard={props.shouldFocusCard}
         setEditing={setEditing}
         setPendingAnnotation={setPendingAnnotation}
+        setAnnotationChangesPending={props.setAnnotationChangesPending}
+        onColorChange={onColorChange}
       />
+
+      {/* Save/Cancel Buttons (only shown when editing existing annotation) */}
       {editingAnnotation && props.data && (
         <ButtonGroup>
           <SaveButton
@@ -232,12 +371,16 @@ function ActiveEditCard({
             saveAnnotation={saveAnnotation}
           />
           <CancelButton
-            props={props}
+            isActive={props.isActive}
             setEditing={setEditing}
             resetAnnotation={resetAnnotation}
+            setAnnotationChangesPending={props.setAnnotationChangesPending}
+            onCancel={props.onCancel}
           />
         </ButtonGroup>
       )}
+
+      {/* Delete Confirmation Dialog */}
       {confirmingDelete && props.data && (
         <Confirmation
           data-testid='confirm-delete'
@@ -256,258 +399,16 @@ function ActiveEditCard({
   );
 }
 
-function useOnRemove(props: EditCardProps, pendingAnnotation: string) {
-  const onRemove = props.onRemove;
-  const trackDeleteHighlight = useAnalyticsEvent('deleteHighlight');
-  const removeAndTrack = React.useCallback(
-    () => {
-      const data = assertDefined(props.data, 'props.data must be defined');
-
-      onRemove();
-      trackDeleteHighlight(data.color);
-    },
-    [onRemove, props.data, trackDeleteHighlight]
-  );
-
-  return props.data && !props.data.annotation && !pendingAnnotation && props.data.color ? removeAndTrack : null;
-}
-
-function AnnotationEditor({
-  props,
-  pendingAnnotation,
-  onColorChange,
-  setEditing,
-  setPendingAnnotation,
-}: {
-  props: EditCardProps;
-  pendingAnnotation: string;
-  onColorChange: ReturnType<typeof useOnColorChange>;
-  setEditing: React.Dispatch<React.SetStateAction<boolean>>;
-  setPendingAnnotation: (value: React.SetStateAction<string>) => void;
-}) {
-  const textarea = React.useRef<HTMLTextAreaElement>(null);
-  const setAnnotationChangesPending = props.setAnnotationChangesPending;
-  const updateUnsavedHighlightStatus = React.useCallback(
-    (newValue: string) => {
-      const currentValue = props.data?.annotation ?? '';
-
-      setPendingAnnotation(newValue);
-      if (currentValue !== newValue && !props.hasUnsavedHighlight) {
-        setAnnotationChangesPending(true);
-      }
-
-      if (currentValue === newValue && props.hasUnsavedHighlight) {
-        setAnnotationChangesPending(false);
-      }
-      setEditing(true);
-    },
-    [
-      props.data?.annotation,
-      props.hasUnsavedHighlight,
-      setAnnotationChangesPending,
-      setEditing,
-      setPendingAnnotation,
-    ]
-  );
-  const initializeColor = React.useCallback(() => {
-    if (!props.highlight.getStyle()) {
-      textarea.current?.blur();
-      onColorChange(highlightStyles[0].label, true);
-      const setFocus = () => {
-        textarea.current?.focus();
-      };
-
-      setFocus();
-      defer(setFocus);
-    }
-  }, [onColorChange, props.highlight]);
-
-  useFocusElement(textarea, props.shouldFocusCard);
-
-  return (
-    <Note
-      textareaRef={textarea}
-      note={pendingAnnotation}
-      onFocus={initializeColor}
-      onChange={updateUnsavedHighlightStatus}
-      edit={Boolean(props.data?.annotation)}
-    />
-  );
-}
-
-function SaveButton({
-  data,
-  setEditing,
-  pendingAnnotation,
-  setConfirmingDelete,
-  saveAnnotation,
-}: {
-  data: HighlightData;
-  setEditing: React.Dispatch<React.SetStateAction<boolean>>;
-  pendingAnnotation: string;
-  setConfirmingDelete: React.Dispatch<React.SetStateAction<boolean>>;
-  saveAnnotation: ReturnType<typeof useSaveAnnotation>;
-}) {
-  const doSave = React.useCallback(
-    (e: React.FormEvent) => {
-      e.preventDefault();
-      setEditing(false);
-      if (pendingAnnotation === '' && data.annotation) {
-        setConfirmingDelete(true);
-      } else {
-        saveAnnotation(data);
-      }
-    },
-    [data, pendingAnnotation, saveAnnotation, setConfirmingDelete, setEditing]
-  );
-
-  return (
-    <FormattedMessage id='i18n:highlighting:button:save'>
-      {msg => (
-        <Button
-          data-testid='save'
-          data-analytics-label='save'
-          size='small'
-          variant='primary'
-          onClick={doSave}
-        >
-          {msg}
-        </Button>
-      )}
-    </FormattedMessage>
-  );
-}
-
-function CancelButton({
-  props,
-  setEditing,
-  resetAnnotation,
-}: {
-  props: EditCardProps;
-  setEditing: React.Dispatch<React.SetStateAction<boolean>>;
-  resetAnnotation: () => void;
-}) {
-  const setAnnotationChangesPending = props.setAnnotationChangesPending;
-  const onCancel = props.onCancel;
-  const cancelEditing = React.useCallback(
-    (e?: React.FormEvent) => {
-      e?.preventDefault();
-      resetAnnotation();
-      setAnnotationChangesPending(false);
-      setEditing(false);
-      onCancel();
-    },
-    [resetAnnotation, setAnnotationChangesPending, setEditing, onCancel]
-  );
-
-  useOnEsc(props.isActive, cancelEditing);
-
-  return (
-    <FormattedMessage id='i18n:highlighting:button:cancel'>
-      {msg => (
-        <Button
-          size='small'
-          data-analytics-label='cancel'
-          data-testid='cancel'
-          onClick={cancelEditing}
-        >
-          {msg}
-        </Button>
-      )}
-    </FormattedMessage>
-  );
-}
-
-function useOnColorChange(props: EditCardProps) {
-  const { highlight, data, locationFilterId, pageId } = props;
-  const trackEditNoteColor = useAnalyticsEvent('editNoteColor');
-  const onCreate = props.onCreate;
-  const dispatch = useDispatch();
-
-  return React.useCallback(
-    (color: HighlightColorEnum, isDefault?: boolean) => {
-      highlight.setStyle(color);
-      if (data) {
-        const { updatePayload, preUpdateData } = generateUpdatePayload(data, {
-          color,
-          id: data.id,
-        });
-
-        dispatch(
-          updateHighlight(updatePayload, {
-            locationFilterId,
-            pageId,
-            preUpdateData,
-          })
-        );
-        trackEditNoteColor(color);
-      } else {
-        assertWindow()
-          .getSelection()
-          ?.removeAllRanges();
-        onCreate(isDefault === true);
-      }
-    },
-    [
-      highlight,
-      data,
-      dispatch,
-      locationFilterId,
-      pageId,
-      trackEditNoteColor,
-      onCreate,
-    ]
-  );
-}
-
-function useSaveAnnotation(
-  { data, pageId, locationFilterId, highlight, onCancel }: EditCardProps,
-  element: React.RefObject<HTMLElement>,
-  pendingAnnotation: string
-) {
-  const dispatch = useDispatch();
-  const trackEditAnnotation = useAnalyticsEvent('editAnnotation');
-
-  return React.useCallback(
-    (toSave: HighlightData) => {
-      const definedData = assertDefined(
-        data,
-        'Can\'t update highlight that doesn\'t exist'
-      );
-
-      const addedNote = definedData.annotation === undefined;
-      const { updatePayload, preUpdateData } = generateUpdatePayload(definedData, {
-        id: toSave.id,
-        annotation: pendingAnnotation,
-      });
-
-      dispatch(
-        updateHighlight(updatePayload, {
-          locationFilterId,
-          pageId,
-          preUpdateData,
-        })
-      );
-      trackEditAnnotation(addedNote, toSave.color);
-      onCancel();
-      scrollHighlightIntoView(highlight, element);
-      // Focus the highlight after save is successful
-      highlight.focus();
-    },
-    [
-      data,
-      dispatch,
-      element,
-      highlight,
-      locationFilterId,
-      onCancel,
-      pageId,
-      pendingAnnotation,
-      trackEditAnnotation,
-    ]
-  );
-}
-
+/**
+ * Styled EditCard component
+ *
+ * Applies visual styling to the card:
+ * - Background color from theme
+ * - Prevents text selection on card controls
+ * - Spacing for button group
+ * - Hides card on touch devices (mobile)
+ */
+// tslint:disable-next-line
 export default styled(EditCard)`
   background: ${theme.color.neutral.formBackground};
   user-select: none;
